requirements.txt
streamlit
langchain
langchain-core
langchain-community
openai
python-dotenv